+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Require Dart `2.12`.
 - Enable null safety.
+- Removed deprecated `createProxyHandler`.
 
 ## 0.1.0+7